Re-style keyword filters
