+++ src/common/components/filterButton/filterButton.module.scss
@@ -3,8 +3,8 @@
 .filter {
   height: auto;
   min-height: 1.5rem;
-  background-color: $hel-metro-light-50;
-  color: $hel-black;
+  background-color: $hel-black-70;
+  color: $hel-white;
   margin: 0.1875rem 0.375rem 0.1875rem 0;
   display: inline-flex;
   padding: 0.125rem 0.5rem;
@@ -20,7 +20,7 @@
     cursor: pointer;
 
     svg {
-      fill: $hel-black;
+      fill: $hel-white;
       height: 0.625rem;
       width: 0.625rem;
       margin: 0;
@@ -31,30 +31,25 @@
 .date,
 .dateType {
   background-color: $hel-engel-light-50;
-}
-.district {
-  background-color: $hel-metro;
-  color: $hel-white;
+  color: $hel-black;
 
   .closeButton {
     svg {
-      fill: $hel-white;
+      fill: $hel-black;
     }
   }
 }
-.keyword {
-  background-color: $hel-metro-light-50;
-}
-.place {
-  background-color: $hel-tram;
-  color: $hel-white;
 
-  .closeButton {
-    svg {
-      fill: $hel-white;
-    }
-  }
-}
+.district,
+.keyword,
+.place,
 .yso {
-  background-color: $hel-metro-light-20;
+  background-color: $hel-black-70;
+  color: $hel-white;
+
+  .closeButton {
+    svg {
+      fill: $hel-white;
+    }
+  }
 }
